requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
requests==2.26.0
schedule==1.1.0
python-dotenv==0.19.0
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
